update: re-organise the files into respective folders
