+++ generative_ai/video/test_video_examples.py
@@ -11,15 +11,21 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-import gemini_describe_http_image_example
-import gemini_describe_http_pdf_example
+import gemini_describe_http_video_example
+import gemini_youtube_video_key_moments_example
+import gemini_youtube_video_summarization_example
 
 
-def test_gemini_describe_http_image_example() -> None:
-    text = gemini_describe_http_image_example.generate_content()
+def test_gemini_describe_http_video_example() -> None:
+    text = gemini_describe_http_video_example.generate_content()
     assert len(text) > 0
 
 
-def test_gemini_describe_http_pdf_example() -> None:
-    text = gemini_describe_http_pdf_example.generate_content()
+def test_gemini_youtube_video_key_moments_example() -> None:
+    text = gemini_youtube_video_key_moments_example.generate_content()
+    assert len(text) > 0
+
+
+def test_gemini_youtube_video_summarization_example() -> None:
+    text = gemini_youtube_video_summarization_example.generate_content()
     assert len(text) > 0
